feat: update v2 sponsored brands (#179)
Updated sponsored brands documentation to match v2.

Tried `oneOf` in line 1420. Doesn't work in the editors render, but I
wonder if it works in mintilfy. The idea is that we show three different
scenarios for the three different triggers. This is to avoid showing all
three and creating confusions (you can only use one at the time).

---------

Co-authored-by: Catalina Elzo <catalina@topsort.com>